+++ tiger-test-lib/src/main/java/io/cucumber/junit/TigerCucumberRunner.java
@@ -12,15 +12,17 @@ import io.cucumber.core.plugin.TigerSerenityReporterPlugin;
 import io.cucumber.core.resource.ClassLoaders;
 import io.cucumber.core.runtime.*;
 import io.cucumber.core.runtime.Runtime;
+import io.cucumber.plugin.Plugin;
 import java.time.Clock;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.LinkedHashSet;
 import java.util.UUID;
 import java.util.function.Supplier;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import net.serenitybdd.core.di.SerenityInfrastructure;
-import net.thucydides.model.webdriver.Configuration;
+import org.apache.commons.lang3.reflect.FieldUtils;
 import org.junit.runners.model.InitializationError;
 
 @Slf4j
@@ -62,18 +64,14 @@ public class TigerCucumberRunner extends CucumberSerenityBaseRunner {
 
   public static Runtime using(
       Supplier<ClassLoader> classLoaderSupplier, RuntimeOptions runtimeOptions) {
-    Configuration<?> systemConfiguration = SerenityInfrastructure.getConfiguration();
-    return createTigerSerenityEnabledRuntime(
-        classLoaderSupplier, runtimeOptions, systemConfiguration);
+    return createTigerSerenityEnabledRuntime(classLoaderSupplier, runtimeOptions);
   }
 
   public static Runtime createTigerSerenityEnabledRuntime(
-      /*ResourceLoader resourceLoader,*/
-      Supplier<ClassLoader> classLoaderSupplier,
-      RuntimeOptions runtimeOptions,
-      Configuration<?> systemConfiguration) {
+      Supplier<ClassLoader> classLoaderSupplier, RuntimeOptions runtimeOptions) {
 
     RuntimeOptionsBuilder runtimeOptionsBuilder = new RuntimeOptionsBuilder();
+
     Collection<String> allTagFilters = environmentSpecifiedTags(runtimeOptions.getTagExpressions());
 
     for (String tagFilter : allTagFilters) {
@@ -81,6 +79,7 @@ public class TigerCucumberRunner extends CucumberSerenityBaseRunner {
     }
 
     runtimeOptionsBuilder.build(runtimeOptions);
+    prependPluginOption(runtimeOptions, TigerSerenityReporterPlugin.class);
     setRuntimeOptions(runtimeOptions);
 
     EventBus bus = new TimeServiceEventBus(Clock.systemUTC(), UUID::randomUUID);
@@ -88,13 +87,10 @@ public class TigerCucumberRunner extends CucumberSerenityBaseRunner {
     FeatureSupplier featureSupplier =
         new FeaturePathFeatureSupplier(classLoaderSupplier, runtimeOptions, parser);
 
-    TigerSerenityReporterPlugin reporter = new TigerSerenityReporterPlugin(systemConfiguration);
-
     Runtime runtime =
         Runtime.builder()
             .withClassLoader(classLoaderSupplier)
             .withRuntimeOptions(runtimeOptions)
-            .withAdditionalPlugins(reporter)
             .withEventBus(bus)
             .withFeatureSupplier(featureSupplier)
             .build();
@@ -109,6 +105,9 @@ public class TigerCucumberRunner extends CucumberSerenityBaseRunner {
     Assertions.assertNoCucumberAnnotatedMethods(clazz);
 
     RuntimeOptions runtimeOptions = createRuntimeOptions(clazz);
+
+    prependPluginOption(runtimeOptions, TigerSerenityReporterPlugin.class);
+
     log.info("Tag filters {}", runtimeOptions.getTagExpressions());
     JUnitOptions junitOptions = createJUnitOptions(clazz);
     initializeBus();
@@ -123,12 +122,28 @@ public class TigerCucumberRunner extends CucumberSerenityBaseRunner {
 
     ThreadLocalRunnerSupplier runnerSupplier = initializeServices(clazz, runtimeOptions);
 
-    Configuration<?> systemConfiguration = SerenityInfrastructure.getConfiguration();
-    TigerSerenityReporterPlugin reporter = new TigerSerenityReporterPlugin(systemConfiguration);
-    addPlugin(reporter);
-
     initiateContext(exitStatus, runnerSupplier);
 
     createFeatureRunners(getFeatures(), runtimeOptions, junitOptions);
+  }
+
+  @Override
+  @SneakyThrows
+  protected void initiateContext(ExitStatus exitStatus, ThreadLocalRunnerSupplier runnerSupplier) {
+    RunnerSupplier tigerRunnerSupplier = new TigerRunnerSupplier(runnerSupplier);
+    FieldUtils.writeField(
+        this,
+        "context",
+        new CucumberExecutionContext(getEventBus(), exitStatus, tigerRunnerSupplier),
+        true);
+  }
+
+  @SneakyThrows
+  private static void prependPluginOption(
+      RuntimeOptions runtimeOptions, Class<? extends Plugin> pluginClass) {
+    var plugins = runtimeOptions.plugins();
+    plugins.add(0, PluginOption.forClass(pluginClass));
+
+    FieldUtils.writeField(runtimeOptions, "plugins", new LinkedHashSet<>(plugins), true);
   }
 }